requirements.txt
asgiref==3.6.0
astroid==2.15.5
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-datetime-utc==1.0.4
django-environ==0.10.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
psycopg2==2.9.6
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
typing_extensions==4.6.0
urllib3==2.0.2
whitenoise==6.4.0
wrapt==1.15.0
